Add files via upload
This code calculates the average patient pay for a given medication for a diagnosis under one insurance grouping(Bin, PCN, Group). This is the average a patient would be expected to pay if they don't have access to the formulary status of the drug, and are chosing a medication at random. The code then calculates the difference between the calculated average, and the price of the highest ranked formulary status, showing an average savings. The highest savings from a picking a drug treatment plan without knowing formulary status vs knowing formulary status is $241.55 for the patient's out of pocket cost. The average savings of picking a drug treatment plan without knowing formulary status vs knowing formulary status is $20.50 for the patient's out of pocket cost.